Fixed No Instance For SpaceTime Schedule
Since was using tail instead of last, was making a variable a
[MultipleFiringOp] instead of a MultipleFiringOp (list vs singleton).
Call space or time on that didn't work as [MultipleFiringOp] is not
an instance of SpaceTime. By default, Haskell would try to make
Schedule instance of SpaceTime work, that would fail, and would say
that space/time function for SpaceTime Schedule wasn't getting a schedule
input. It should say no SpaceTime instance for [MultipleFiringOp], but I
got not instance for.